back-end/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
black==24.10.0
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distlib==0.3.9
distro==1.9.0
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.1.0
django-cors-headers==4.5.0
django-health-check==3.18.3
django-stubs==5.1.0
django-stubs-ext==5.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
elastic-transport==8.17.1
elasticsearch==8.17.1
filelock==3.18.0
frozenlist==1.6.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.10
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.4
langchain-core==0.3.61
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langsmith==0.1.147
multidict==6.4.4
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==1.26.4
openai==1.82.0
orjson==3.10.18
packaging==24.2
pathspec==0.12.1
pgvector==0.3.6
platformdirs==4.3.8
pre_commit==4.0.1
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
types-PyYAML==6.0.12.20250516
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.2
yarl==1.20.0
